grades threshold changes

+++ BTfaculty/models.py
@@ -17,7 +17,7 @@ class BTGradesThreshold(models.Model):
     Grade = models.ForeignKey('BTsuperintendent.BTGradePoints', on_delete=models.CASCADE)
     Subject = models.ForeignKey('BTco_ordinator.BTSubjects', on_delete=models.CASCADE)
     RegEventId = models.ForeignKey('ADUGDB.BTRegistrationStatus', on_delete=models.CASCADE)
-    Threshold_Mark = models.IntegerField()
+    Threshold_Mark = models.FloatField()
     Section = models.CharField(max_length=2, default='NA')
     Exam_Mode = models.BooleanField() 
     class Meta:
